change some values to decimal, start calculation method

+++ BucParking/Models/ParkingSpot.cs
@@ -5,8 +5,8 @@ namespace BucParking.Models
     public class ParkingSpot
     {
         public string ParkingLotId { get; set; } = string.Empty;
-        public double Latitude { get; set; }
-        public double Longitude { get; set; }
+        public decimal Latitude { get; set; }
+        public decimal Longitude { get; set; }
         public ParkingType Type { get; set; }
 
         public static ParkingSpot SpotsFromCsv(string csvLine)
@@ -14,8 +14,8 @@ namespace BucParking.Models
             String[] values = csvLine.Split(',');
             ParkingSpot parkingSpotsData = new ParkingSpot();
             parkingSpotsData.ParkingLotId = (values[5]);
-            parkingSpotsData.Latitude = double.Parse(values[1]);
-            parkingSpotsData.Longitude = double.Parse(values[0]);
+            parkingSpotsData.Latitude = decimal.Parse(values[1]);
+            parkingSpotsData.Longitude = decimal.Parse(values[0]);
             parkingSpotsData.Type = Enum.Parse<ParkingType>(values[6], true);
 
             return parkingSpotsData;
